# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses": {
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new course",
        "tags": [
          "Courses"
        ]
      },
      "get": {
        "operationId": "CoursesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all courses",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get a course by id",
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update a course",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete a course",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/certificates": {
      "get": {
        "operationId": "CertificatesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all certificates",
        "tags": [
          "Certificates"
        ]
      },
      "post": {
        "operationId": "CertificatesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Issue a new certificate",
        "tags": [
          "Certificates"
        ]
      }
    },
    "/certificates/{id}": {
      "get": {
        "operationId": "CertificatesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get certificate by ID",
        "tags": [
          "Certificates"
        ]
      },
      "patch": {
        "operationId": "CertificatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update certificate status",
        "tags": [
          "Certificates"
        ]
      },
      "delete": {
        "operationId": "CertificatesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Revoke certificate",
        "tags": [
          "Certificates"
        ]
      }
    },
    "/disputes": {
      "get": {
        "operationId": "DisputesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all disputes",
        "tags": [
          "Disputes"
        ]
      },
      "post": {
        "operationId": "DisputesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Raise a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/disputes/stats": {
      "get": {
        "operationId": "DisputesController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get dispute aggregate statistics",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/disputes/{id}": {
      "get": {
        "operationId": "DisputesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get dispute by ID",
        "tags": [
          "Disputes"
        ]
      },
      "patch": {
        "operationId": "DisputesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update dispute fields",
        "tags": [
          "Disputes"
        ]
      },
      "delete": {
        "operationId": "DisputesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Remove dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/disputes/{id}/resolve": {
      "post": {
        "operationId": "DisputesController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Resolve a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/disputes/{id}/reject": {
      "post": {
        "operationId": "DisputesController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Reject a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/disputes/{id}/comments": {
      "post": {
        "operationId": "DisputesController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Add a comment or admin note to a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/disputes/{id}/priority": {
      "patch": {
        "operationId": "DisputesController_updatePriority",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update the priority of a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/enrollments": {
      "get": {
        "operationId": "EnrollmentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all enrollments",
        "tags": [
          "Enrollments"
        ]
      },
      "post": {
        "operationId": "EnrollmentsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create new enrollment",
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/enrollments/{id}": {
      "get": {
        "operationId": "EnrollmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get enrollment by ID",
        "tags": [
          "Enrollments"
        ]
      },
      "patch": {
        "operationId": "EnrollmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update enrollment progress",
        "tags": [
          "Enrollments"
        ]
      },
      "delete": {
        "operationId": "EnrollmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Cancel enrollment",
        "tags": [
          "Enrollments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coursera App API",
    "description": "The API for Admin and Instructor modules",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the course",
            "example": "Introduction to Programming"
          },
          "subtitle": {
            "type": "string",
            "description": "Subtitle of the course"
          },
          "description": {
            "type": "string",
            "description": "Description of the course"
          },
          "instructor": {
            "type": "string",
            "description": "The name of the instructor",
            "example": "Dr. Sarah Jenkins"
          },
          "organization": {
            "type": "string",
            "description": "The organization",
            "example": "University of Tech"
          },
          "category": {
            "type": "string",
            "description": "The category",
            "example": "Computer Science"
          },
          "level": {
            "type": "string",
            "description": "The level of the course",
            "example": "Advanced"
          },
          "price": {
            "type": "number",
            "description": "The price of the course",
            "example": 49.99
          },
          "students": {
            "type": "number",
            "description": "Number of students",
            "example": 1250
          },
          "rating": {
            "type": "number",
            "description": "Rating",
            "example": 4.8
          },
          "status": {
            "type": "string",
            "description": "The status",
            "example": "Published"
          },
          "image": {
            "type": "string",
            "description": "Thumbnail image URL"
          },
          "objectives": {
            "type": "object",
            "description": "Learning objectives"
          },
          "prerequisites": {
            "type": "string",
            "description": "Prerequisites"
          },
          "modules": {
            "description": "Modules and curriculum",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "type": "boolean",
            "description": "Visibility"
          },
          "certificate": {
            "type": "boolean",
            "description": "Certificate"
          },
          "createdDate": {
            "type": "string",
            "description": "Created date",
            "example": "2023-05-10"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the course",
            "example": "Introduction to Programming"
          },
          "subtitle": {
            "type": "string",
            "description": "Subtitle of the course"
          },
          "description": {
            "type": "string",
            "description": "Description of the course"
          },
          "instructor": {
            "type": "string",
            "description": "The name of the instructor",
            "example": "Dr. Sarah Jenkins"
          },
          "organization": {
            "type": "string",
            "description": "The organization",
            "example": "University of Tech"
          },
          "category": {
            "type": "string",
            "description": "The category",
            "example": "Computer Science"
          },
          "level": {
            "type": "string",
            "description": "The level of the course",
            "example": "Advanced"
          },
          "price": {
            "type": "number",
            "description": "The price of the course",
            "example": 49.99
          },
          "students": {
            "type": "number",
            "description": "Number of students",
            "example": 1250
          },
          "rating": {
            "type": "number",
            "description": "Rating",
            "example": 4.8
          },
          "status": {
            "type": "string",
            "description": "The status",
            "example": "Published"
          },
          "image": {
            "type": "string",
            "description": "Thumbnail image URL"
          },
          "objectives": {
            "type": "object",
            "description": "Learning objectives"
          },
          "prerequisites": {
            "type": "string",
            "description": "Prerequisites"
          },
          "modules": {
            "description": "Modules and curriculum",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "type": "boolean",
            "description": "Visibility"
          },
          "certificate": {
            "type": "boolean",
            "description": "Certificate"
          },
          "createdDate": {
            "type": "string",
            "description": "Created date",
            "example": "2023-05-10"
          }
        }
      }
    }
  }
}